\name{Fluvicolines}
\alias{Fluvicolines}
\docType{data}
\title{
The Fluvicolines data set
}
\description{
Phylogeny and distributions for the Fluvicoline clade of New World Flycatchers
}
\usage{data(Fluvicolines)}
\format{
A \code{nodiv_data} object with the phylogeny and distributions of all species
}

\source{
Ohlson, J., Irestedt, M., Ericson, P.G. & Fjeldså, J. (2013). Phylogeny and classification of the New World suboscines (Aves, Passeriformes). Zootaxa, 3613, 1–35. 
Rahbek, C., Hansen, L. A. & Fjeldså, J. (2012). One degree resolution database for the global distribution of birds. 
}

\examples{
data(Fluvicolines)
}
\keyword{datasets}
