\name{update_object}
\alias{update_object}

\title{
Update \code{distrib_data} objects to the \code{nodiv} version >= 1.1 format
}
\description{
Creates the \code{species_stats} element
}
\usage{
update_object(distrib_data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{distrib_data}{
An object of type \code{distrib_data}, \code{nodiv_data} or \code{nodiv_result}
}
 }

\value{
An updated version of the object
}

\author{
Michael Krabbe Borregaard
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
