\name{Fluvicolines}
\alias{Fluvicolines}
\docType{data}
\title{
The Fluvicolines data set
}
\description{
Phylogeny and distributions for the Fluvicoline clade of New World Flycatchers
}
\usage{data(Fluvicolines)}
\format{
a \code{nodiv_data} object with the phylogeny and distributions of all species
}
\source{
Ohlson, J., Irested, M., Ericson, P.G. and Fjeldsaa, J. (2013). Phylogeny and classification of the New World suboscines (Aves, Passeriformes). Zootaxa, 3613, 1-35., AND Rahbek, C., Hansen, L.A. and Fjeldss, J. (2012). One degree resolution database for the global distribution of birds.
}
\examples{
data(Fluvicolines)
plot(Fluvicolines)
}
\keyword{datasets}